update messages and chats domain entity

+++ src/Deepin.Application/DTOs/Chats/ChatUnreadCount.cs
@@ -1,12 +1,10 @@
-using System;
-
 namespace Deepin.Application.DTOs.Chats;
 
 public class ChatUnreadCount
 {
     public Guid ChatId { get; set; }
-    public int UnreadCount { get; set; }
-    public ChatUnreadCount(Guid chatId, int unreadCount)
+    public long UnreadCount { get; set; }
+    public ChatUnreadCount(Guid chatId, long unreadCount)
     {
         ChatId = chatId;
         UnreadCount = unreadCount;
